refactor: cast uv calculations to int for accurate indexing in diffuse method

+++ model.cpp
@@ -111,7 +111,7 @@ TGAColor Model::diffuse(Vec2i uv)
 Vec2i Model::uv(int iface, int nvert)
 {
     int idx = faces_[iface][nvert].get(1,0);
-    return Vec2i{{uv_[idx].get(0,0) * diffusemap_.get_width()}, {uv_[idx].get(1,0) * diffusemap_.get_height()}};
+    return Vec2i{{static_cast<int>(uv_[idx].get(0,0) * diffusemap_.get_width())}, {static_cast<int>(uv_[idx].get(1,0) * diffusemap_.get_height())}};
 }
 
 Vec3f Model::GetNorm(int iface, int nvert)
